Query::set_layout: setting the layout on the subarray.
When setting the layout on a query, we did not set the layout on the
subarray. In python, the code will then go and call the size estimation
API  and cache the range offsets (range_offsets_) with the wrong layout.
When running the query later, these are not computed again and in some
cases will create this segfault.

---
TYPE: IMPROVEMENT
DESC: Query::set_layout: setting the layout on the subarray.