Introduce malloc_trim on Linux (#2443)
TileDB currently has some deficiencies in how we alloc and free
internal memory. There is short and medium term efforts to improve this,
however malloc_trim provides a (potentially) large benefit to users on
Linux until improvements are finalized.

The main issue is TileDB allocs, and reallocs buffers that can be very
small in size while tiles are read and run through the filter pipeline,
along with other operations such as schema reading, metadata reading,
etc. The potentially large number of small allocations or reallocation
can cause fragmentation and can also cause the arena allocator to no
unmap memory that has been freed. This happens in a most obvious way on
Linux with glibc allocation in that a long running process might see
memory usage creep upwards after every query is finished but not
necessarily immediately drop on query completion or context destruction.
This is because while the process is running the allocator might not
unmap the memory pages due to the usage patterns.

The short term solution is we introduce calls to malloc_trim on context
destruction and on query class destruction for linux only.