Allow empty attribute writes. (#2461)
The error condition hit when trying to set an empty attribute has now
been removed and a small fix to the global order writer was made. This
allowed to successfully write and read empty attributes. Moreover,
looking at the unit test that validated this error, it referenced
https://github.com/TileDB-Inc/TileDB/issues/590 which showed code that
originally segfaulted and caused the addition of the error condition.
This code, with the removal of the error condition, does not segfault
anymore. A valid attribute file is also successfully written to disk.

---
TYPE: IMPROVEMENT
DESC: Allow empty attribute writes.