Allow null data pointers for writes. (#2481)
As a minor follow up to #2476 we expand the empty attribute writes to allow the data buffer to be null. This is useful for C++ users who might be using `std::vector` for their underlying data buffer where the vector's `std::vector::data()` will return `nullptr` when the vector is empty.

---
TYPE: IMPROVEMENT
DESC: Allow null data pointers for writes.