Array Schema needs checked and load per fragment
The array schema is used to determine some details in the fragment
footers. While this is not ideal, it is how the current code behaves. As
such we need to load the schema name first and then set the proper array
schema on each fragment.

We also need to make sure to generate unique URIs on schema evolution
invocation.