Fix support for returning fill-in value with new attributes
This properly supports new attributes added from schema evolution
returning the fill-in value as the default when older fragments are
included in a query. Minor adjustments are made to the read algorithm to
support the cases where an attribute or dimension's tile does not exist
due to the attribute not existing in the older fragment.