Refactored readers can segfault when multiple contexts are used. (#2525)
* Refactored readers can segfault when multiple contexts are used.

The OpenArrayMemoryTracker lives in the OpenArray objects owned by the
storage manager. If a query doesn't receive the same context as the one
that was used to open the array, we weren't able to retrieve the
OpenArrayMemoryTracker object. This uses the global state to retreive
the object from all the existing storage manager objects. This is a
work around to allow VCF queries to work for now. Moving forward, we
should prevent a query to have a different context than the opened
array.

---
TYPE: IMPROVEMENT
DESC: Refactored readers can segfault when multiple contexts are used.

* Rename array_get_memory_tracker to array_memory_tracker

Co-authored-by: Seth Shelnutt <seth@tiledb.io>