Reset the size of scratch buffer for each REST query
This fixes a bug in which the scratch buffer used for the partial curl
call back did not reset it size after processing. After we complete
processing one query we copy the "left over" data which is part of the
next incomplete query (usually 1-2KB of curl data only) to the start of
the scratch buffer. This lets us reuse the scratch buffer and not have
to resize or realloc it. We need to reset the "valid" byte size to only
the copied amount.