Add regression test for sc-11082
Test that the query execution with empty result does not write out
of buffer bounds. We create an oversize buffer with guard values
ahead of the actual pointer range given to libtiledb; we run the
query to completion with empty result; then we check the guard
values. This test fails prior to PR#2540.