WIP: Wrap all c-api in a function-try-block
This prevents any exceptions from escaping. Previously we had attempted
to wrap all c++ calls in try/catch blocks. This however is prune to
mistakes and missing parts of the code that might throw an exception.
Using the function-try-block guarantees that all functions will have
their exceptions caught.

For simplicity of the PR all existing function definitions are
unchanged. As such there is duplicate try/catching which can be cleaned
up in followup PRs.