Refactored dense reader: fixing output buffer offsets with multi-ranges.
When a partition includes multiple-ranges, the destination of the cell
slabs was only computed properly for the first range. For the subsequent
ranges, we need to compute how many cells were already in the previous
ranges and start to copy data at that point. We also need to use the
correct range start coordinates to compute the correct offset inside
of that particular subarray for the copy.

---
TYPE: IMPROVEMENT
DESC: Refactored dense reader: fixing output buffer offsets with multi-ranges.