Add new TileDB File API
This adds a new TileDB file API which allows the storing of arbitrary
files as a 1 dimensional dense array. The fact is all files are arrays.
We can easily store any file as an array and gain the benefits of TileDB
arrays. This includes versioning/time traveling, remote object store
support, compressors/filters and even partial file reading!

The support here is preliminary, there is multiple improvements left on
the table for followups. This includes improved support for partial file
reading, exporting, and batch style ingestion.

A new dependency on libmagic is introduced inorder to detect details
about a file, MIME, encoding, compression, etc.