Refactored sparse readers: serialization fixes. (#2558)
This fixes serialization for queries using the refactored readers. First
the client didn't know when the server was done with a query as it
didn't know if the server had finished processing result tiles. This is
now properly serialized. Second, as the server or client might not have
the fragment metadata loaded when the query is deserialized, it couldn't
calculate the memory usage of the loaded result tiles, this is now
computed only once the query is resumed. It also properly sets the
memory budget on the server when doing deserialization. Finally, the
array uri used to be serialized but this was causing issues and as the
array is opened both on the client and server, we stop sending it back
and forth.

---
TYPE: IMPROVEMENT
DESC: Refactored sparse readers: serialization fixes.

Co-authored-by: Seth Shelnutt <Shelnutt2@gmail.com>