Query object now owns its REST scratch space (#2555)
This is an improvement that optimizes REST requests. When a query
overflows the user buffer we might have the next part of the query
already loaded into the scratch buffer. As such we should avoid
discarding it and instead keep the lifetime the same as the query
itself. This optimizes the user's next call to submit where the user
resumes the incomplete.