Fixing pre-loading for tile offsets in various readers.
Investigating the deadlock trying to switch to the new threadpool, I
found out that read_tiles was trying to load tile offsets while holding
a lock, which is too much work. The reason this happened is because tile
offsets, which are supposed to be preloaded, were not being cached.
as the subarray used to determine the relevant fragments didn't have
relevant fragments computed. This is now fixed by using the root
subarray inside of the partitioner which has relevant fragments computed.

Also, to prevent this class of problems moving forward, this change is
changing the fragment metadata apis that retrieve tile offsets to error
out when tile offsets are not loaded instead of going to load them in
the background. All tile offsets should be cached before calling those
apis from now on.

This also includes an optimization in compute_result_space_tiles to only
look at relevant fragments.

---
TYPE: IMPROVEMENT
DESC: Fixing pre-loading for tile offsets in various readers.