Convert loose files thread_pool.* into a unit (#2520)
* Convert loose files thread_pool.* into a unit.

This change uses a proxy header for `thread_pool.h`, a change that can be considered transitional. There are open questions about how to specify common headers. They might require explicit paths in #include directives, or specifying include paths in the build, or some other mechanism, or perhaps even to stick with proxy headers in a common location.

* Add `unit_thread_pool` to `tests` target

* clang-format