Sparse unordered with dups reader: removing result cell slabs.
Cell slabs generation has proven to be a bottle neck for some customers.
The main issue here is that the generation was single threaded. This
change is moving from using cell slabs to cache tile bitmaps, when there
is a subarray set so that the proper cells can be moved from the tiles
to the user’s buffers. The copy operation is now run in parallel across
all result tiles. Query condition is now applied directly on the
bitmaps.

Also included in this change is:
- A simplification of the memory budget for both the sparse global order
reader and the sparse unordered with duplicates reader by folding in the
result tiles budget into the coords budget.
- Moving the sparse global order reader to the same bitmap caching and
query condition processing so that cell slabs improvements can be made.
- Calculation of the query condition tiles budget before processing for
the sparse global order reader so there is no more failures in query
condition. (shortcut item #9401). This also simplifies some of the
common tile processing functions greatly.

---
TYPE: IMPROVEMENT
DESC: Sparse unordered with dups reader: removing result cell slabs.