Add OptWithDebInfo build type and LIBRARY_CMAKE_BUILD_TYPE var.
This change adds a new build type, `OptWithDebInfo`, which is equivalent
to `RelWithDebInfo`, except that it continues to build all libraries in
`Release` mode, unless overridden by an explicit
`LIBRARY_CMAKE_BUILD_TYPE` setting.

Libraries that previously used the `CMAKE_BUILD_TYPE` to choose their
build type are now migrated to use the `LIBRARY_CMAKE_BUILD_TYPE`,
which is set to the same as `CMAKE_BUILD_TYPE`, except in the case of
`OptWithDebInfo`, when it is set to `Release`.

This is intended to support creating a fully `-O3` optimized build
while still including DWARF symbols for better debug info from the cloud
server running in production.

The binaries compiled before and after these changes are byte-identical
for all the default types (`Debug`, `Release`, `RelWithDebInfo`, and
`OptWithDebInfo`), and after stripping, the `Release` and
`OptWithDebInfo` types are also byte-identical (with the exception
of a handful of bytes in the header):

```diff
41,42c41,42
< 00000280: 9c37 e09a d127 846d 4832 ff33 3c73 c1e0  .7...'.mH2.3<s..
< 00000290: b672 4c7d 0000 0000 0902 0000 fa01 0000  .rL}............
---
> 00000280: f616 35f6 0208 ce3d 6eb1 e7ab 36ab 2454  ..5....=n...6.$T
> 00000290: 5781 0a9c 0000 0000 0902 0000 fa01 0000  W...............
```