Sparse unordered with duplicates reader: cell num fix. (#2636)
ResultTile::cell_num() only uses the tile from the first dimension to
calculate the number of cells. In the sparse unordered with duplicates
reader, we don't load all coordinate tiles so there are cases where
the first dimension might not be loaded and the user requests only the
second dimension. This will return a cell number of 0 and make the
reader segfault.

---
TYPE: IMPROVEMENT
DESC: Sparse unordered with duplicates reader: cell num fix.