Read_tiles parallelization improvements.
This changes read tiles to parallelize on both attributes and tiles
when generating the list of work for VFS. Also, in the past, the
function held a lock on result_tiles_mutex_ while doing all memory
allocations for reading a particular attribute. This now changes it to
only lock around a few data structure calls, which speeds up the
allocation process quite a bit.

---
TYPE: IMPROVEMENT
DESC: Read_tiles parallelization improvements.