Handle query condition causing incomplete query
If the query has a condition that filters out a tile, the tile needs to
be removed from the query. The bug is when the tile remains but all
cells have been filtered the query results in incomplete state since a
tile remains but each submit skips the tile (since cell num is 0).