Fix handling checking config for refactored reader
This supports the deprecated parameters for "sm.use_refactored_readers",
which is needed for serialization. If the client is pre 2.5, then they
(in nearly all cases) want to use the legacy reader not the new
refactored reader. This adjusts the check for refactored or not to
account for the 2.4 config parameter which defaulted to false. The 2.4
config parameter was replaced by finer grained config settings in 2.5.