Set the array uri for cap'n proto compatiblity
In 2.4 and older unfortunately the client didn't check to see if the
cap'n proto object had the array uri set and instead always assume it
was and force set the value on the Array object. This meant setting it
to an empty string, effectively removing the URI from the array.

In 2.5 we need to capture the sent URI, and stash it temporarily to
return it to the client on the cap'n proto array object.