Sparse unordered with duplicates: Better vectorization for tile bitmaps.
When computing tile bitmaps, there are a few places where we were not
vectorizing properly and it caused performance impacts on some customer
scenario. Fixing the use of covered/overlap in the dimension class to
enable proper vectorization.

---
TYPE: IMPROVEMENT
DESC: Sparse unordered with duplicates: Better vectorization for tile bitmaps.