Sparse refactored reader: change all_tiles_loaded_ to vector of uint8_t.
Investigating random failures in unit tests for the sparse global order
reader, I found out the all_tiles_loaded_ was sometimes incorrect. This
was happening because the backing container was a vector of bools, which
internally gets converted to a bitset. As we use this member variable in
parallel for all fragments, it was sometimes leading to inpredictable
results. Switching to a vector of uint8_t ensures that a full byte is
used for all fragments and fixes the parallelisation issues.

---
TYPE: IMPROVEMENT
DESC: Sparse refactored reader: change all_tiles_loaded_ to vector of uint8_t.