Sparse refactored readers: Better vectorization for tile bitmaps calculations. (#2711)
* Sparse unordered with duplicates: Better vectorization for tile bitmaps.

When computing tile bitmaps, there are a few places where we were not
vectorizing properly and it caused performance impacts on some customer
scenario. Fixing the use of covered/overlap in the dimension class to
enable proper vectorization.

---
TYPE: IMPROVEMENT
DESC: Sparse unordered with duplicates: Better vectorization for tile bitmaps.

* Adding better vectorization for compute_results_count_sparse.

* Switching to stack allocations for some small vectors.

* Remove bitmap from overlap_vec usage

This is an optimization to avoid a double loop over all ranges by
inserting to the relevant indexes directly.

* Code cleanup.

* Moving the covered bitmap from a resource pool to the stack.

* Moving unfiltered buffer allocation closer to the filtered buffer
allocation to speed up the filter pipeline.

* WIP: remove attr_tiles_mtx

Co-authored-by: Seth Shelnutt <seth@tiledb.io>