Sort ranges for unordered with duplicate reader and exit comparisons early (#2736)
Sort ranges for unordered with duplicate reader and exit comparisons early

This adds the ability to sort the subarray ranges on a query. The sorted ranges are then used by the sparse unordered with duplicates reader to facilitate the tile MBR and cell comparisons. When there is a large number of ranges (100k+) this can reduce the time spent in the bitmap computations by allowing less comparisons against the ranges for a given dimension.