Avoid setting config on transient subarray (#2740)
For the existing `tiledb_query_add_range` APIs there is a transient
subarray object that is referenced. In one API the query config was also
being set on this subarray. This is not required, and other variations
of `tiledb_query_add_range` did not set the config. Removing this saves
a few seconds of the WALL time when a user is setting hundreds of
thousands of ranges.

The performance bottleneck comes in when the config is copied which is a
full copy of the config map. This was being done once per range setting.

Instead to maintain current behavior setting the query config also sets
the subarray config.