Sparse unordered w/ dups: var buffer overflow on tile continuation fix.
In the sparse unordered with duplicates reader, when we continue a query
on a partially processed tile and we hit an overflow of the var sized
buffer, we didn't take into account the minimum cell position when
resolving the overflow, which caused undefined behavior.

---
TYPE: IMPROVEMENT
DESC: Sparse unordered w/ dups: var buffer overflow on tile continuation fix.