Sparse unordered w/ dups reader: remove invalid assert.
After we are done processing tile bitmaps and query condition, we remove
tiles that didn't make the cut. It's totally possible that every tiles
that fit into memory for a call to dowork all get filtered out, so this
assert needs to be removed as it is invalid.

---
TYPE: IMPROVEMENT
DESC: Sparse unordered w/ dups reader: remove invalid assert.