Improve object_type detection performance
This improves the object APIs performance for detecting types by
switching from listing all items in the URI to checking only for the
existence of the group indicator, array schema file or array schema
folder. We also switch the order to check for the array schema folder
first, since it is most likely to exist based on the assumption that
there are more arrays than there are groups.