Query condition: differentiate between nullptr and empty string.
While porting the existing tests for query condition to the refactored
readers, I found out that empty string and nullptr were treated the same
when initializing a query condition, as initializing a vector to size 0
doesn't guarantee an allocation, and differentiation was done on the
vector::data() member.

Also included in this change:
- Better vectorization for the refactored dense reader.
- Ported existing query condition tests to the refactored dense and sparse readers.
- Some code cleanup.

---
TYPE: IMPROVEMENT
DESC: Query condition: differentiate between nullptr and empty string.