Support storing integral cells in a chunk.
Currently chunks are broken up at 64kib boundary for unfiltered data.
This means that potentially we'll have a single cells value split across
two chunks. This is not ideal as we then have to included multiple
chunks to access a single cell.

This change makes sure cells are never split. The heuristic is the
following when determining to add a cell to a chunk:
 - If the cell fits in the buffer, add it.
 - If it doesn't fit and new size < 150% capacity, add it.
 - If it doesn't fit and current size < 50% capacity, add it.

---
TYPE: IMPROVEMENT
DESC: Support storing integral cells in a chunk.