Removing Buffer from Tile.
The Buffer class is used everywhere in the system, and because of that,
it's trying to do too much. This is a first step to remove some of those
uses by removing Buffer entirely from Tile. Tile should be a simple
container that manages a unfiltered and filtered buffer.

Also, the ultimate goals of tiles is that they are allocated with a size
and that will not be changed. For the reader, it's very simple to know
up front what the size of a tile is. So for the read path in this
change, the unfiltered size of tiles are all defined in 'read_tiles'
through the use of 'alloc_data'. Filtered sizes are also defined in
'read_tiles'. This will allow soon to make it so that tiles are
constructed from the get go with a size.

For the writer, it's a little more challenging, but still achievable.
This change is a step in the right direction, fixed size attributes are
allocated and not changed. Var size attributes will take some work. For
now, 'WriterTile' exposes a write_var method that does realloc when the
size of the buffer is exceeded. In the future this will need to change
to compute sizes of var sized tiles first.

So, the change also cleans up a lot of the tile usage. The tile no
longer manages an offset and so every read/writes now specifies which
offset the operation is happening on.

---
TYPE: IMPROVEMENT
DESC: Removing Buffer from Tile.