Sparse unordered w dups reader: fixing query continuation with subarray. (#2824)
This change fixes calculations for the case where a query with subarray
was incomplete, and only one tile is left, partially processed. The
cells copied calculations for this tile was incorrect and the tile was
determined to be incomplete, which would send the reader into an
infinite loop.

---
TYPE: IMPROVEMENT
DESC: Sparse unordered w dups reader: fixing query continuation with subarray.