Dense reader: fixing src cell offset for different cell order.
In the refactored dense reader, when the user queries in a cell order
that is different than the array cell order, there was an issue where
the src cell in the tile was not computed properly. This addresses the
issue.

---
TYPE: IMPROVEMENT
DESC: Dense reader: fixing src cell offset for different cell order.