Fixing Subarray::crop_to_tile to crop default dimensions. (#2831)
The refactored reader is using Subarray::crop_to_tile to process every
tiles independently. When setting a range on only one dimension for a
2D array, this API wouldn't crop the other dimension as it is a default
one. This caused the refactored dense reader to try to copy data that is
outside of the tile currently processed, causing an out of bound read
for a tile. Since this API is only available to fixed data types, all
calls will have the default range filled in, so it doesn't need to skip
default dimensions when processing.

---
TYPE: IMPROVEMENT
DESC: Fixing Subarray::crop_to_tile to crop default dimensions.