Sparse unordered w dups reader: fixing max pos calculation in tile copy. (#2840)
* Sparse unordered w dups reader: fixing max pos calculation in tile copy.

Continuing on #2824, there was still an issue with the calculation for
tile continuation in copy tiles. In dev, this was masked because
pos_with_given_result_sum was changed to allow asking for a  result sum
that was out of bound and return the max position of the tile.

Also in dev, there is still a bug for the following corner case: a
query ends on the last tile, but cannot copy it. Continuation goes and
the last partial tile still cannot fit in the user buffers and needs to
be split again. This will give incorrect results.

In 2.6, this would return max uint64 for the position and would crash.

---
TYPE: IMPROVEMENT
DESC: Sparse unordered w dups reader: fixing max pos calculation in tile copy.

* Fixing an issue with serialized queries.