Adding min/max/sum/null count to fragment info.
This adds the min/max/sum/null count information to the written fragment
info. It also adds the reader side functionality to load and read the
min/max/sum/null count for a specific attribute. This will later enable
query condition to do faster filtering on the data and enable
aggregate functionalities.

The processing of the min/max/sum/null count for a tile is done by the
TileMetadataGenerator class, which is easily unit testable.

Implementing this change, a new WriterTile class was implemented on top
of the existing Tile class. This is so that the added per tile metadata
by the Writer doesn't affect the size of the Tile objects for the
readers. One of the existing member, pre_filtered_size_, was moved to
this new class as it's only used by the writer.

---
TYPE: IMPROVEMENT
DESC: Adding min/max/sum/null count to fragment info.