Writer: processing var tiles before offset tiles.
To store integral cells in a tile, the filter pipeline needs the offsets
tile while processing the var tile. As the filter pipeline frees the
data buffer of a processed tile, this introduced an issue where in some
cases, the offsets buffer would be freed as the var tile is being
processed. This would be a strong case for using a smart pointer, but
unfortunately, this change is quite extensive. This change just makes
sure that var tiles are always processed before offset tiles.

---
TYPE: IMPROVEMENT
DESC: Writer: processing var tiles before offset tiles.