Global writes: writting bad validity values when coordinates are dups. (#2848)
While working on another PR, I noticed an if statement with no brackets
is causing a bug in the writer. This will write invalid validity tiles
when there are duplicates coordinates and we are writting in global
order by the validity cell of duplicate coordinates when it shouldn't.
It will only happen on fixed size attributes.

---
TYPE: IMPROVEMENT
DESC: Global writes: writting bad validity values when coordinates are dups.