Writer: moving unordered_map::emplace outside of parallel_for.
In Writer::ordered_write, there is an unordered_map::emplace inside of
a parallel_for, which is not safe. As the call just emplaces an empty
vector for an attribute, it's better to just do it outside of the
parallel_for for all attributes.

---
TYPE: IMPROVEMENT
DESC: Writer: moving unordered_map::emplace outside of parallel_for.