Sparse unordered w/ dups reader: all empty string attribute fix. (#2874)
In the sparse unordered with duplicates reader, the size of the output
var size buffer was used to determine if any cell can fit in the buffers.
Unfortunately, if all cells are empty, the size will be 0. This fix
changes the condition check to use the computed cell offsets instead.

---
TYPE: BUG
DESC: Sparse unordered w/ dups reader: all empty string attribute fix.