Tile metadata test: reducing amount of spew in verbose mode.
There has been several CI failures on the tile metadata tests with no
crash dump. The test is asking for a lot of metadata that is not present
on purpose to test that this codepath works and since it's doing so on
a lot of variations, this causes a lot of spew when running in verbose
mode. Looking more closely at the logs, I noticed the following clue:

awk: cmd. line:1: (FILENAME=- FNR=8894) fatal: print to "standard output" failed (Resource temporarily unavailable)

This could hint that something is unstable in the standard output when
running tests in github actions, so reducing the amount of spew seems
like a valid attempt to reduce the amount of CI failures so this change
does just that.

TYPE: IMPROVEMENT
DESC: Tile metadata test: reducing amount of spew in verbose mode.