Reader: treating empty string range as expected.
In a lot of places in the code, an empry string range means default.
This leads to a lot of confusion as the user can also pass in empty
string ranges and an empty string is a valid coordinate. This change
removes the assumption that empty strings are default in a few
functions in the dimension class and relies only on std::string_view
compare. They are:
- Dimension::covered<char>
- Dimension::overlap<char>
- Dimension::overlap_ratio<char>
- Dimension::relevant_ranges<char>
- Dimension::covered_vec<char>

To fix the default range behavior for strings, two functions needed to
be adjusted at the subarray level: compute_relevant_fragments and
compute_relevant_fragment_tile_overlap. For the first one, the fix was
trivial, simply default the relevant fragment bitmap to 1 for default
ranges and skip the computation. For the tile overlap however, it meant
that the is_default values need to be passed in through the rtree to
the domain as the overlap ratio computation for all dimensions is done
at that level.

---
TYPE: IMPROVEMENT
DESC: Reader: treating empty string range as expected.