compute_results_count_sparse_string: fixing incorect memcmp. (#2892)
This was recently found while running VCF queries. There is an issue in
compute_results_count_sparse_string where code to check if a vector
contains only zeros. The code was copied from compute_results_sparse<char>
and as it was adapted to work for both tile bitmaps (uint8_t) and tile
count bitmaps, two issues were introduced.

The first one was the computation of the memory buffer size, where
sizeof(uint64_t) was used to multiply the number of elements, which
works for count bitmaps but  could cause a segfault for regular bitmaps.

The second one would affect count bitmaps only. It is using a optimized
way to check if all elements of a buffer are 0, which is to use
addr[0] == 0 && memcmp(addr[0], addr[1], size - 1). This works fine if
addr is a char pointer but not if it is a wider byte size value, which
for count buffers was uint64_t. So the second part of the fix casting
to pass the proper addr + 1 in.

---
TYPE: IMPROVEMENT
DESC: compute_results_count_sparse_string: fixing incorect memcmp.