Sparse rindex readers: fixing query resume on TileDB cloud. (#2900)
There are two types of incompletes on TileDB Cloud. One where the user's
buffers are not full, but the server side buffers are and one where the
user side buffers are full. In the first case, the query is 'kept alive'
with it's current state. In the second the query is recreated on the
server and all state is lost. This fixes an issue in the second case
where the all ranges tile overlap gets recomputed when the query is
resumed. In this case, the computation needs to take into account the
tile index deserialized from the client.

---
TYPE: IMPROVEMENT
DESC: Sparse rindex readers: fixing query resume on TileDB cloud.