Listing improvements: new directory structure for array. (#2918)
* Listing improvements: new directory structure for array.

When opening an array, certain remote file systems limit the number of
elements returned (for example: 1000 for s3). For arrays with a large
number of fragments can take a long time. In an effort to reduce the
size of the listing operations, this PR shuffles a few elements of the
array directory.

First of all, it moves all the ok files to a __commit folder and the
fragments to a __fragment folder.

Second, it moves the consolidated fragment metadata to a __fragment_meta
directory.

Finally, this PR also ensures that arrays with the old format, the new
format or a mix of both can be read.

---
TYPE: IMPROVEMENT
DESC: Listing improvements: new directory structure for array.

* Addressing feedback from @stavrospapadopoulos.