Sparse unordered w/ dups reader: fixing memory management for tiles.
This fixes a few issues with the reader. First, it doesn't clear query
condition tiles during tile copy as they should be treated as coordinates
tiles.

Second, when the result tiles vector needs to be shrunk because the var
size data cannot fit the user's buffers, we now clear the result tiles
from memory right away for all loaded attributes. This solves an issue
where some tiles would be lingering in memory.

Third, it fixes computation of the offsets when a query continues to
fill in the user's buffers after running out of memory budget for tiles.

---
TYPE: IMPROVEMENT
DESC: Sparse unordered w/ dups reader: fixing memory management for tiles.