Support reading all consolidated fragment metadata files.
This change fixes an issue with fragment metadata where the latest
timestamp cannot always be trusted. To fix this, all consolidated
fragment are loaded.

Vacuuming is still using the latest timestamp but now considers the
latest timestamp on every URI so multiple fragment metadata might be
left around. This doesn't match the behavior of the WRT consolidation,
where we only vacuum the consolidated files if a complete set of the
fragments is found inside of a consolidated file. As for now the
vacuuming API doesn't take in the key, we cannot know if a particular
file contains everything. This can be addressed in a later PR.

---
TYPE: IMPROVEMENT
DESC: Support reading all consolidated fragment metadata files