Global writes: fixing OOM on write continuation.
convert_tile_min_max_var_sizes_to_offsets is used to convert a buffer of
sizes to offsets. For global writes, the function will be called
multiple times on the same buffer so the function should resume from the
last tile that was processed. This information is already saved
in tile_index_base_ so the fix is trivial.

---
TYPE: IMPROVEMENT
DESC: Global writes: fixing OOM on write continuation.