Dense reader: fix user buffer offset computation for multi-index queries.
The previous implementation of the dense reader would order the results
in the user buffer by N-dimension range index, for example, if there are
2 ranges set per dimensions, the data would be ordered from N-dimension
range 1 to 4. This implementation is incorrect as it doesn't really
return the data in the expected row/column major.

This change computes the correct offset for the data in the user buffers
and returns the data in true row/column major order.

---
TYPE: IMPROVEMENT
DESC: Dense reader: fix user buffer offset computation for multi-index queries.