Improve Group Support (#2966)
Improve Group Support

This PR is the first (and likely largest) in a series of PRs that will
improve the TileD Group Support.

This PR focuses on adding a new on-disk format in which we explicitly
keep track of members of a group. New c-apis/cpp are added to support
this. Groups now can be opened in read and write mode. Arrays or groups
can be added as members via `tiledb_group_add_member` and
`tiledb_group_remove_member`.

Metadata on groups is also now supported. It uses the same basic
functionality as arrays to support groups.

Co-authored-by: bdeng-xt <bdeng.xt@gmail.com>
Co-authored-by: Isaiah Norton <isaiah@tiledb.io>