Support consolidating non-contiguous fragments.
This adds support to consolidate non-contiguous fragments as an
experimental API. I f the user is not careful, for dense, this could
lead to data loss if the joint fragment domains of the consolidated
fragments overlap fragments domains of in between fragments. For sparse,
it could lead to ordering errors if there is overlap of data of the
fragments in between the consolidated fragments. This will later be
fixed by adding a system time variable to consolidated fragments.

---
TYPE: IMPROVEMENT
DESC: Support consolidating non-contiguous fragments.