Sparse unordered w/ dups reader: coord tiles management fix. (#3023)
When clearing tiles during a var size overflow, we should only clear
attribute tiles as coord tiles are counted in another memory budget and
cleaned up later. There was a small logic bug where is_dim was used to
determine if we were dealing with a dimension or not, but this variable
was for the attribute currently being copied, and not the attribute
considered for cleanup.

---
TYPE: IMPROVEMENT
DESC: Sparse unordered w/ dups reader: coord tiles management fix.