Dense reader: do not sort input ranges.
Following up on https://github.com/TileDB-Inc/TileDB/pull/3002, it was
determined we should not sort the input ranges. If the user has a dense
vector of domain 1-10 and requests ranges 3-4 then 1-2, we should
return the ranges in that order.

---
TYPE: IMPROVEMENT
DESC: Dense reader: do not sort input ranges.