Global writer: fixing write continuation for fixed sized attributes.
When continuing global writes with a partially written tile, the partial
tile is completed and swapped with the first tile in the result vector,
then the remainder of the cells is written but should start at the next
tile, which was not the case in the current implementation. This fixes
the issue.

Note that this doesn't include unit tests to unblock the 2.8 release.

---
TYPE: IMPROVEMENT
DESC: Global writer: fixing write continuation for fixed sized attributes.