All ranges tile overlap: skip computation for default dimensions.
When computing tile overlap for the refactored sparse readers, default
dimensions were not taken into consideration. For fixed size data, it
was not an issue as default values get filled in, but for string
dimension, it would get the tile overlap for an empty range, which in
the end is "" - "", which will return no data. The solution is just to
skip processing for default dimensions and take into account the default
flag when generating the final tile overlap structure.

---
TYPE: IMPROVEMENT
DESC: All ranges tile overlap: skip computation for default dimensions.