Sparse unordered w/ dups reader: fix incomplete reason for cloud reads.
Status incomplete reason was using the fact that for deserialized
queries, result tiles were not created to know it was a cloud read. The
reader changed and this behavior is no longer true. This fixes the
issue by using is_remote() on the array, which is the proper way.

---
TYPE: IMPROVEMENT
DESC: Sparse unordered w/ dups reader: fix incomplete reason for cloud reads.