In cell/tile_order_cmp use const ref to dimension
These are extremely hot paths and incrementing the atomic int for
reference counting can cause a significant (measured at >2x) in these
paths. We solve this by switching to a constant reference to the
shared_ptr to avoid incrementing the reference counter. These paths are
completely controlled inside the Domain itself, so there is not lifetime
concerns by using the reference.