Global writes: check global order on write continuation. (#3109)
* Global writes: check global order on write continuation.

When doing write continuation, we used to only validate global order on
the cells for the current write. This allowed for subsequent writes to
not necessarily respect global order. This adds a check to make sure the
first cell of a continuation write comes after (in either global or
hilbert order) the last cell written by a previous write.

Also fixing sc#16777.

---
TYPE: IMPROVEMENT
DESC: Global writes: check global order on write continuation.

* Fix wrong error condition (last_tile_cell_idx == 0).

* Addressing feedback from @ihnorton.