Consolidation w/ timestamps: writing timestamps tiles. (#3155)
* Consolidation w/ timestamps: writing timestamps tiles.

This modifies the sparse global order reader to fill in the __timestamps
buffer with the timestamp data. It also modifies read_tiles and
unfilter_tiles to skip tiles with no timestamp data, and allows the tile
copy for the reader to copy timestamps values depending on wether or not
the tile contains timestamps information (using the fragment timestamp
if it doesn't). Finally, this modifies the merge operation to include
all duplicated coordinates into the cells slabs for consolidation reads,
but it does not, for regular reads select the appropriate cell (the one
with the greater timestamp) in consolidated fragments with timestamps.
This will be added in a later PR.

---
TYPE: IMPROVEMENT
DESC: Consolidation w/ timestamps: writing timestamps tiles.

* Addressing feedback from @ypatia.

* Addressing feedback from @ypatia, part 2.